feat: confirm when delete user

+++ src/components/Admin/Listuser/Listuser.tsx
@@ -17,16 +17,19 @@ const Listuser = () => {
     }, []); // Dependency array rỗng để useEffect chỉ chạy một lần sau khi component mount
     // Hàm xử lý sự kiện khi nhấn nút xóa người dùng
     const handleDeleteUser = (userId) => {
-        // Gọi API để xóa người dùng với id tương ứng
-        axios.delete(`http://localhost:3000/googleAccount/${userId}`)
-            .then((response) => {
-                console.log('Xóa người dùng thành công:', response.data);
-                // Cập nhật lại danh sách người dùng sau khi xóa thành công
-                setUserData((prevUserData) => prevUserData.filter((user) => user.id !== userId));
-            })
-            .catch((error) => {
-                console.error('Lỗi khi xóa người dùng:', error);
-            });
+        const shouldDelete = window.confirm('Bạn có chắc chắn muốn xóa người dùng này?');
+        if (shouldDelete) {
+            // Gọi API để xóa người dùng với id tương ứng
+            axios.delete(`http://localhost:3000/googleAccount/${userId}`)
+                .then((response) => {
+                    console.log('Xóa người dùng thành công:', response.data);
+                    // Cập nhật lại danh sách người dùng sau khi xóa thành công
+                    setUserData((prevUserData) => prevUserData.filter((user) => user.id !== userId));
+                })
+                .catch((error) => {
+                    console.error('Lỗi khi xóa người dùng:', error);
+                });
+        }
     };
     return (
         <div className="listproduct">
@@ -48,7 +51,7 @@ const Listuser = () => {
                             <td>{user.email}</td>
                             <td><img src={user.img} alt="" /></td>
                             <td>
-                                <button style={{padding:10, backgroundColor:'red',border:'none',color:'white',borderRadius:5}} onClick={() => handleDeleteUser(user.id)}>Xóa</button>
+                                <button style={{ padding: 10, backgroundColor: 'red', border: 'none', color: 'white', borderRadius: 5 }} onClick={() => handleDeleteUser(user.id)}>Xóa</button>
                             </td>
                         </tr>
                     ))}
